+++ subprojects/dependency-management/src/main/java/org/gradle/api/internal/artifacts/configurations/DefaultVariant.java
@@ -32,7 +32,10 @@ import org.gradle.api.internal.collections.DomainObjectCollectionFactory;
 import org.gradle.api.internal.file.FileCollectionFactory;
 import org.gradle.internal.Describables;
 import org.gradle.internal.DisplayName;
+import org.gradle.internal.Factory;
 import org.gradle.internal.typeconversion.NotationParser;
+
+import java.util.List;
 
 public class DefaultVariant implements ConfigurationVariantInternal {
     private final Describable parentDisplayName;
@@ -40,6 +43,7 @@ public class DefaultVariant implements ConfigurationVariantInternal {
     private AttributeContainerInternal attributes;
     private final NotationParser<Object, ConfigurablePublishArtifact> artifactNotationParser;
     private final PublishArtifactSet artifacts;
+    private Factory<List<PublishArtifact>> lazyArtifacts;
 
     public DefaultVariant(Describable parentDisplayName, String name,
                           AttributeContainerInternal parentAttributes,
@@ -60,7 +64,7 @@ public class DefaultVariant implements ConfigurationVariantInternal {
     }
 
     public OutgoingVariant convertToOutgoingVariant() {
-        return new LeafOutgoingVariant(getAsDescribable(), attributes, artifacts);
+        return new LeafOutgoingVariant(getAsDescribable(), attributes, getArtifacts());
     }
 
     private DisplayName getAsDescribable() {
@@ -80,6 +84,10 @@ public class DefaultVariant implements ConfigurationVariantInternal {
 
     @Override
     public PublishArtifactSet getArtifacts() {
+        if (lazyArtifacts != null) {
+            artifacts.addAll(lazyArtifacts.create());
+            lazyArtifacts = null;
+        }
         return artifacts;
     }
 
@@ -98,6 +106,11 @@ public class DefaultVariant implements ConfigurationVariantInternal {
     @Override
     public String toString() {
         return getAsDescribable().getDisplayName();
+    }
+
+    @Override
+    public void artifactsProvider(Factory<List<PublishArtifact>> artifacts) {
+        this.lazyArtifacts = artifacts;
     }
 
     @Override
