+++ subprojects/plugins/src/main/java/org/gradle/api/plugins/GroovyBasePlugin.java
@@ -21,7 +21,6 @@ import org.gradle.api.Plugin;
 import org.gradle.api.Project;
 import org.gradle.api.Task;
 import org.gradle.api.Transformer;
-import org.gradle.api.artifacts.Configuration;
 import org.gradle.api.file.ConfigurableFileCollection;
 import org.gradle.api.file.FileCollection;
 import org.gradle.api.file.FileTreeElement;
@@ -30,8 +29,7 @@ import org.gradle.api.internal.plugins.DslObject;
 import org.gradle.api.internal.tasks.DefaultGroovySourceSet;
 import org.gradle.api.internal.tasks.DefaultSourceSet;
 import org.gradle.api.model.ObjectFactory;
-import org.gradle.api.plugins.internal.JavaPluginsHelper;
-import org.gradle.api.plugins.internal.SourceSetUtil;
+import org.gradle.api.plugins.internal.JvmPluginsHelper;
 import org.gradle.api.provider.Provider;
 import org.gradle.api.reporting.ReportingExtension;
 import org.gradle.api.specs.Spec;
@@ -40,7 +38,6 @@ import org.gradle.api.tasks.SourceSet;
 import org.gradle.api.tasks.compile.CompileOptions;
 import org.gradle.api.tasks.compile.GroovyCompile;
 import org.gradle.api.tasks.javadoc.Groovydoc;
-import org.gradle.internal.component.external.model.TestFixturesSupport;
 
 import javax.inject.Inject;
 import java.io.File;
@@ -115,31 +112,18 @@ public class GroovyBasePlugin implements Plugin<Project> {
                 final Provider<GroovyCompile> compileTask = project.getTasks().register(sourceSet.getCompileTaskName("groovy"), GroovyCompile.class, new Action<GroovyCompile>() {
                     @Override
                     public void execute(final GroovyCompile compile) {
-                        SourceSetUtil.configureForSourceSet(sourceSet, groovySourceSet.getGroovy(), compile, compile.getOptions(), project);
+                        JvmPluginsHelper.configureForSourceSet(sourceSet, groovySourceSet.getGroovy(), compile, compile.getOptions(), project);
                         compile.dependsOn(sourceSet.getCompileJavaTaskName());
                         compile.setDescription("Compiles the " + sourceSet.getName() + " Groovy source.");
                         compile.setSource(groovySourceSet.getGroovy());
                     }
                 });
-                SourceSetUtil.configureOutputDirectoryForSourceSet(sourceSet, groovySourceSet.getGroovy(), project, compileTask, compileTask.map(new Transformer<CompileOptions, GroovyCompile>() {
+                JvmPluginsHelper.configureOutputDirectoryForSourceSet(sourceSet, groovySourceSet.getGroovy(), project, compileTask, compileTask.map(new Transformer<CompileOptions, GroovyCompile>() {
                     @Override
                     public CompileOptions transform(GroovyCompile groovyCompile) {
                         return groovyCompile.getOptions();
                     }
                 }));
-
-                if (SourceSet.MAIN_SOURCE_SET_NAME.equals(sourceSet.getName()) || TestFixturesSupport.TEST_FIXTURE_SOURCESET_NAME.equals(sourceSet.getName())) {
-                    // The user has chosen to use the java-library plugin;
-                    // add a classes variant for groovy's main sourceSet compilation,
-                    // so default-configuration project dependencies will see groovy's output.
-                    project.getPluginManager().withPlugin("java-library", new Action<AppliedPlugin>() {
-                        @Override
-                        public void execute(AppliedPlugin plugin) {
-                            Configuration apiElements = project.getConfigurations().getByName(sourceSet.getApiElementsConfigurationName());
-                            JavaPluginsHelper.registerClassesDirVariant(compileTask, project.getObjects(), apiElements);
-                        }
-                    });
-                }
 
                 // TODO: `classes` should be a little more tied to the classesDirs for a SourceSet so every plugin
                 // doesn't need to do this.
